@@ -1,17 +1,22 @@
-name: example
+name: jaspr_supabase_example
 description: Example usage of jaspr_supabase
 version: 0.0.1
 publish_to: none
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.8.0 <4.0.0'
 
 dependencies:
-  jaspr: ^0.16.1
-  jaspr_supabase:
-    path: ../
-  supabase: ^2.0.0
+  jaspr: ^0.22.1
+  jaspr_supabase: ^26.2.24
+  jaspr_router: ^0.8.1
 
 dev_dependencies:
-  build_runner: ^2.4.0
-  jaspr_builder: ^0.16.1
+  build_runner: ^2.9.0
+  build_web_compilers: ^4.4.7
+  jaspr_builder: ^0.22.2
+  jaspr_lints: ^0.6.0
+  lints: ^5.0.0
+
+jaspr:
+  mode: client
